moving categories around
