+++ libs/light/docs/reference/light/rainbow-cycle-animation.md
@@ -12,7 +12,7 @@ light.rainbowCycleAnimation();
 ## Example
 
 ```blocks
-control.forever(() => {
+loops.forever(() => {
     light.builtin.showAnimationFrame(light.rainbowCycleAnimation())
 })
 ```
