+++ libs/light/docs/reference/light/running-lights-animation.md
@@ -12,7 +12,7 @@ light.runningLightsAnimation();
 ## Example
 
 ```blocks
-control.forever(() => {
+loops.forever(() => {
     light.builtin.showAnimationFrame(light.runningLightsAnimation())
 })
 ```
