+++ libs/light/docs/reference/light/sparkle-animation.md
@@ -12,7 +12,7 @@ light.sparkleAnimation();
 ## Example
 
 ```blocks
-control.forever(() => {
+loops.forever(() => {
     light.builtin.showAnimationFrame(light.sparkleAnimation())
 })
 ```
