+++ libs/light/test.ts
@@ -7,7 +7,7 @@ function flash(n: number) {
     control.runInBackground(() => {
         strip.setPixelColor(n, 0x0000ff)
         strip.show()
-        control.pause(1000)
+        loops.pause(1000)
         strip.setPixelColor(n, 0x000000)
         strip.show()
     })
