+++ libs/microphone/test.ts
@@ -1,4 +1,4 @@
-control.forever(() => {
+loops.forever(() => {
     let level = input.soundLevel()
     serial.writeValue("sound", level)
 })
