+++ libs/photon/docs/examples/photon-spark.md
@@ -2,7 +2,7 @@
 
 ```blocks
 photon.setMode(PhotonPenMode.PenUp);
-control.forever(() => {
+loops.forever(() => {
     photon.forward(Math.random(10))
     control.pause(20)
 })
